+++ CHANGES.md
@@ -32,6 +32,7 @@
 ### Other
 
 - DOC: Creating a real `base` notebook and renaming the old one to `optical`
+- DOC: Better type hints (replacing `XDS_TYPE` by `xr.DataArray`)
 - CI: Using `sertit.ci.reduce_verbosity` instead of recreating the function
 
 ## 0.13.1 (2022-03-08)